+++ fpaq/fpaq3b.cpp
@@ -10,7 +10,7 @@
 namespace std {}  // for MARS compiler
 using namespace std;
 typedef unsigned char U8;
-typedef unsigned long U32;  // 32 bit type
+typedef unsigned int U32;  // 32 bit type
 #define Top_value U32(0XFFFFFFFF)    /* Largest code value */
 /* HALF AND QUARTER POINTS IN THE CODE VALUE RANGE. */
 #define First_qtr U32(Top_value/4+1)  /* Point after first quarter    */
