+++ pasqda/pasqda30/pasqda30.cpp
@@ -562,7 +562,7 @@ template <class T> inline int size(const T& t) {return t.size();}
 // 8-32 bit unsigned types, adjust as appropriate
 typedef unsigned char U8;
 typedef unsigned short U16;
-typedef unsigned long U32;
+typedef unsigned int U32;
 
 #define Top_value U32(0XFFFFFFFF)	/* Largest code value */
 /* HALF AND QUARTER POINTS IN THE CODE VALUE RANGE. */
